raspberrypi/measure_temp/requirements.txt
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.1
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
